fix repeated sticker from stickerForwards. added ignoreGroups add the group id in that to ignore the sticker from group.

+++ stickerForward.js
@@ -4,12 +4,14 @@ const WSF = require("wa-sticker-formatter");
 const fs = require('fs');
 const getRandom = (ext) => { return `${Math.floor(Math.random() * 10000)}${ext}` };
 
-const forwardGroup = "120363024616188056@g.us"; // Group ID
+const forwardGroup = ""; // Group ID
+
+const ignoreGroup = ["", ""]; // Group ID with commas
+
+const stickerLengthArray = [];
 
 const stickerForward = async (sock, msg, from) => {
-    if (!forwardGroup) return sock.sendMessage(from, { text: "Forward Group not found" }, { quoted: msg });
-
-    if (from == forwardGroup) return;
+    if (!forwardGroup || ignoreGroup.includes(from) || from === forwardGroup) return;
 
     if (msg.message.extendedTextMessage) {
         msg['message'] = msg.message.extendedTextMessage.contextInfo.quotedMessage
@@ -21,22 +23,33 @@ const stickerForward = async (sock, msg, from) => {
     const buffer = await downloadMediaMessage(msg, 'buffer', {});
     fs.writeFileSync(media, buffer);
 
-    try {
-        const webpWithMetadata = await WSF.setMetadata(packName, authorName, media);
-        await sock.sendMessage(forwardGroup, { sticker: Buffer.from(webpWithMetadata) });
-        fs.unlinkSync(media);
-    } catch (e) {
-        const { Sticker, StickerTypes } = require("wa-sticker-formatter-1");
-        const buffer = await new Sticker(media)
-            .setPack(packName)
-            .setAuthor(authorName)
-            .setType(StickerTypes.FULL)
-            .setQuality(80)
-            .toBuffer();
+    fs.stat(media, async (err, stats) => {
+        if (err) return;
+        const fileSizeInBytes = stats.size;
+        const fileSizeInKB = fileSizeInBytes / 1024;
+        const fileSizeInMB = fileSizeInKB / 1024;
+        if (stickerLengthArray.length > 5) stickerLengthArray.shift();
+        if (stickerLengthArray.includes(fileSizeInMB)) return fs.unlinkSync(media);
+        else {
+            try {
+                const webpWithMetadata = await WSF.setMetadata(packName, authorName, media);
+                await sock.sendMessage(forwardGroup, { sticker: Buffer.from(webpWithMetadata) });
+                fs.unlinkSync(media);
+            } catch (e) {
+                const { Sticker, StickerTypes } = require("wa-sticker-formatter-1");
+                const sticker_buffer = await new Sticker(media)
+                    .setPack(packName)
+                    .setAuthor(authorName)
+                    .setType(StickerTypes.FULL)
+                    .setQuality(80)
+                    .toBuffer();
 
-        await sock.sendMessage(forwardGroup, { sticker: buffer });
-        fs.unlinkSync(media);
-    }
+                await sock.sendMessage(forwardGroup, { sticker: sticker_buffer });
+                fs.unlinkSync(media);
+            }
+            stickerLengthArray.push(fileSizeInMB);
+        }
+    });
 }
 
 module.exports = stickerForward;
